Refactor commands in application layer

+++ Pors.Application/Users/Commands/DeleteUserCommand.cs
@@ -1,22 +1,16 @@
 ﻿using System;
 using MediatR;
-using System.Text;
-using System.Linq;
 using System.Threading;
-using FluentValidation;
 using Pors.Domain.Entities;
 using System.Threading.Tasks;
-using System.Collections.Generic;
-using FluentValidation.Validators;
-using Microsoft.EntityFrameworkCore;
-using Pors.Application.Common.Models;
 using Pors.Application.Common.Interfaces;
+using Pors.Application.Common.Exceptions;
 
 namespace Pors.Application.Users.Commands
 {
     #region command
 
-    public class DeleteUserCommand : IRequest<Result>
+    public class DeleteUserCommand : IRequest
     {
         public int Id { get; set; }
     }
@@ -25,31 +19,11 @@ namespace Pors.Application.Users.Commands
 
     #region validation
 
-    public class DeleteUserCommandValidator : AbstractValidator<DeleteUserCommand>
-    {
-        private readonly ISqlDbContext _dbContext;
-
-        public DeleteUserCommandValidator(ISqlDbContext dbContext)
-        {
-            _dbContext = dbContext;
-
-            RuleFor(x => x.Id)
-                .MustAsync(BeUserExist).WithMessage("کاربری مطابق شناسه دریافتی یافت نشد");
-        }
-
-        public async Task<bool> BeUserExist(int userId, CancellationToken cancellationToken)
-        {
-            var result = await _dbContext.Users.AnyAsync(x => x.Id == userId, cancellationToken);
-
-            return result;
-        }
-    }
-
     #endregion;
 
     #region handler
 
-    public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Result>
+    public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand>
     {
         private readonly ISqlDbContext _dbContext;
 
@@ -58,15 +32,20 @@ namespace Pors.Application.Users.Commands
             _dbContext = dbContext;
         }
 
-        public async Task<Result> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+        public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
             var entity = await _dbContext.Users.FindAsync(request.Id);
+
+            if (entity == null)
+            {
+                throw new NotFoundException(nameof(User), request.Id);
+            }
 
             _dbContext.Users.Remove(entity);
 
             await _dbContext.SaveChangesAsync();
 
-            return Result.Success();
+            return Unit.Value;
         }
     }
